spi-zynq-qspi: Fill in mode_bits and remove private MODEBITS check
Defining "spi-cpha" in the device-tree makes the Zynq QSPI driver
fail to initialize with the following message:
m25p80 spi32766.0: setup: unsupported mode bits 3

The spi core driver checks the "mode_bits" field of the master, which
spi-zynq-qspi did not fill in so it is still zero.

Fill in the mode_bits field with the MODEBITS define. Remove the now
duplicate check in zynq_qspi_setup_transfer, because the spi core has
already checked this.

Tested with a Spansion flash that supports both modes 0 and 3.

Signed-off-by: Mike Looijmans <mike.looijmans@topic.nl>