zynq-miami: Remove SDIO1 controller and add VGA/LVDD IP to devicetree
Remove the SDIO1 controller, it interferes and does not work.
Remove the unused clock chip driver (didn't work anyway).
Add the VGA/LVDD display driver IP core to the DT. This uses an
VDMA block to transfer the data (much like the HDMI core). Driver
is in prototype stage and may be added later.