usb/host/zynq-dr-of.c: Support deferral on GPIO controller
If the GPIO controller is not available yet, do not simply ignore this but
defer probe until the controller is available.

This allows the USB reset to be controlled by an I2C GPIO expander. The
I2C bus is probed typically after the USB bus, which causes the reset
line for the ULPI to remain in a "low" state on a board that uses a GPIO
expander to control this line.

Signed-off-by: Mike Looijmans <mike.looijmans@topic.nl>