sdhci: Allow external limit on clock frequency
Clock may be limited by other factors than just the card and controller,
like for example routing and wiring.
The Zynq for example has a 25MHz limit when the SDIO lines are routed
through the FPGA logic.

Add a "max-frequency" property to the devicetree that defines extra
limits. Add a variable named "clk_limit" as the "max" frequency in the
code refers to the frequencies that the controller can generate.

Signed-off-by: Mike Looijmans <mike.looijmans@topic.nl>