zynq-dr-of: Allow use of tri-state reset GPIO
To avoid having to use level converters, the reset GPIO can also be
modelled as a tristate.
When a property "xlnx,phy-reset-gpio-tristate" is set, the reset line
will be tri-stated instead of activated high after probing the driver.
Using a pull-up on the BRESET line, the 2v5 ULPI reset can be controlled
by a 1v8 or 3v3 device by only pulling it to ground.

Signed-off-by: Mike Looijmans <mike.looijmans@topic.nl>