topic-miami.dtsi: Enable second UART
Enable the second UART through EMIO. On the Florida board, this can be
connected to the FTDI chip which provides 4 serial ports, one is connected
to the debug serial port, the others to logic. By simply routing the
second FTDI port to UART1 it can be used as a high-speed serial connection
with flow control, e.g. for a PPP link.