usb/host/zynq-dr-of: Trigger PHY RESET early in probe
Move the (optional) reset of the PHY up in the probe method. This has
the following advantages:
- The chip will be ready for operation sooner.
- When the GPIO is not available yet, the clock start/stop can be avoided.

Reduce the "sleep" after the reset, the PHY chip will keep the DIR pin
asserted until it is ready to receive commands, so there is no need to
wait 1 ms after deasserting reset.

Signed-off-by: Mike Looijmans <mike.looijmans@topic.nl>