dts/topic-miami: Add OTG mode for USB
The USB port will automatically switch between host and peripheral mode
depending on the plug inserted. A micro-B plug will select peripheral
mode and set up a network link with the host on the other side. A micro-A
plug will enable the 5V VBUS power supply and attach a stick or other
gadget. Note that the plug must have the ID pin connected to GND to work
properly in host mode (connect pin 4 and 5), most converter cables do not
have this (they have a micro-B connector).

Tested with a cable that I manually patched to connect pins 4 and 5.