sound/soc/adi/axi-spdif.c: Support programmable master clock
If the master clock supports programmable rates, program it to generate
the desired frequency. Only apply constraints when the clock is fixed.

The clock frequency must be set before enabling it. Enabling the clock
was don in "startup", but that occurs before "hw_params" where the rate
is known. Move the clock start to the hw_params routine, and keep track
of whether the clock has been started, because shutdown may be called
without having called hw_params first.

Starting the clock and enabling the SPDIF output AFTER programming the
dividers is a more logical order anyway.

Signed-off-by: Mike Looijmans <mike.looijmans@topic.nl>