sound/soc/adi/axi-spdif.c: Fix "fixed clock" detection and clock refcounting
The detection did not work, implement it differently:

To detect if the source clock is fixed, the driver calls clk_round_rate
for two frequencies. If the results are equal, or if the call returns
an error, the driver assumes the clock is fixed.

hw_params may be called multiple times, only start the clock if it wasn't
already running.