regulator/fixed.c: Don't report EPROBE_DEFER errors
If the regulator cannot be registered because its supplier is not
available yet, don't write an error. There is no need to alert the
user of probe deferrals.
This fixes a storm of error messages at boot when a GPIO controlled
regulator is supplied by an I2C controlled supply.

Signed-off-by: Mike Looijmans <mike.looijmans@topic.nl>