Add topic-miami-florida-*.dts devicetrees
Split florida boards into "gen", "med" and "pcie" type, there's a lot less
on the PCIe board than on the GEN board. Use a common file for whatever they
have in common. Don't define dyplo, just one dts for that.

Allows the build to create DTS files for all boards, preventing strange
results when the drivers for devices not on the platform are loaded anyway.