MTD: m25p80: fix inconsistency in m25p_ids for Micron devices
As stated in a5b7616c5, "mtd: m25p80,spi-nor: Fix module aliases for
m25p80", m25p_ids[] in m25p80.c needs to be kept in sync with
spi_nor_ids[] in spi-nor.c.

This patch fixes the mismatches for the Micron devices, the
"n25q256a" and "n25q512a" do not exist in the spi_nor_ids, so
replace them with the correct names for these chips.

This repairs the disappearance of NOR flash on the Miami boards since 3.18.

Signed-off-by: Mike Looijmans <mike.looijmans@topic.nl>