usb: chipidea: Reduce ULPI PHY reset pulse to datasheet spec of 1us
The datasheet for the 334x PHY mentions that a reset can be performed:
"... by bringing the pin low for a minimum of 1 microsecond and
then high."
A delay of 5ms to implement that seems overly long, so reduce it to
just 1us.
As for the delay after reset, the datasheet only mentioned that the
chip will assert the DIR output. 1ms seems like a safe time to wait
for that to happen, so no change there.

Signed-off-by: Mike Looijmans <mike.looijmans@topic.nl>