HACK: Set MTU for ethernet gadget to 512
The ChipIdea USB driver appears to be broken. Sending more than 512 bytes
in a single packet will 'hang' the USB driver. As a workaround, set the
default MTU to 512 for ethernet gadgets.

Intended as a temporary workaround, not a permanent fix. Though one could
use this patch to increase the MTU and get much better RNDIS performance.