topic-miami: Use "bitbang" I2C controller instead of the Cadence IP
The "hard" I2C controller in the Zynq appears to be plagued with bugs that
won't be fixed in hardware, and they cannot all be worked around. Symptom
on the Florida boards is that it is hard to get the adv7611 HDMI input chip
to work reliably with this controller. When replaced with the "bitbang"
controller on the same MIO pins, things do work okay.

Note: Also enable GPIO bitbang controller in kernel config for this to work.

Also adds the pinmux for the heartbeat LED, which was used in testing problems
with probing order of the GPIO and pinmux controllers on the Zynq.

see also:
http://www.xilinx.com/support/answers/47916.html
https://ez.analog.com/message/223881#223881