hwmon/ltc2990: Sync with mainline
This is the LTC2990 driver as it will appear in Linux mainline.

Use standard hwmon interface files in sysfs. Userspace clients must
be adapted to match this, as the filenames have changed.
The chip is now configured in continuous sampling mode. This speeds
up reading at the cost of consuming 1mW of power.