Changed clocksource for hdmi_tx
Use the CDCE (programmable clock chip) on the Florida to provide
the clock for HDMI-TX. This CDCE chip is ideal to generate clock
frequencies for video stuff because of its 27MHz base crystal.

Before the clock from the axi_clkgen from ADI was used, so this
device is removed from the devicetree.