topic-miamiplus.dtsi: Fix typo "max6650". Rename si514. I2C1 pinmux.
Mistaken ID, not 6550 but 6650 for the max6650 chip.

Name the si514 on the Miami plus boards "si514miami" to prevent name
collisions with carriers boards that already have an si514.

The I2C1 bus moved from 52+53 to 48+49 to make room for the MDIO.
Adjust the pinmux registers to accomodate for that change (and
prevent the bus being muxed to two pins).