zynq_remoteproc: Support reserved-memory via devicetree
The firmware for the remote CPU is to be loaded into a fixed memory location.
A proper way to create a memory hole like that is to specify it as
reserved-memory in the devicetree.

Change zynq_remoteproc to support a "memory-region" link to a reserved
memory range. To preserve backward compatibility the driver will fall back
to using the "reg" entry if the memory-region phandle was not found. Add
a working example to the devicetree binding documentation. This makes the
memory hole definition in the devicetree consistent with what the driver
will use for loading the firmware, and avoids having to use cmdline "mem"
overrides.