dma: xilinx_dma: Only enable VDMA frame IRQ when useful
By default, the VDMA engine sends an interrupt when a frame is ready.
If the DMA client did not register a callback for the descriptor,
the interrupt handler will do nothing at that point, and the CPU
wakes up for nothing.

Arm the frame and delay interrupts only when the client registered
a callback.

Signed-off-by: Mike Looijmans <mike.looijmans@topic.nl>