ethernet/cadence/macb.c: Don't try to set a 0 tx_clk rate
When clocking the Zynq GEM TX from programmable logic, the clock
driver does not know what frequency it is running, so it returns
a zero on clk_round_rate. Instead of complaining and using this
value, silently ignore the clock rate and assume that the clock
can take care of itself.

Signed-off-by: Mike Looijmans <mike.looijmans@topic.nl>