spi-nor: Fix spi-nor read and write in edge cases
spi_nor_read handle odd read requests by doing an additional
small alignment read. spi_nor_write solves it the same way and
it resolves the issue of an out of bound write when length is
smaller then 2. It also prevents a kmalloc.

spi_nor_read For chips with multi die support the spi controller
has to split the read in chunks of 16MB. A multiply of 2 was
missing in the remaining page length calculation which resulted
in data corruption.

spi_nor_write The remaining bank length used in case of
addr_width 4 in stacked configuration was missing a multiply which
is there in the read code

Signed-off-by: Leon Leijssen <leon.leijssen@topic.nl>