media: axi-hdmi-rx: Initialize bus_type
The v4l2-fwnode expects the bus_type member passed to v4l2_fwnode_endpoint_parse
to have an actual value. Set it to V4L2_MBUS_UNKNOWN (0) to request the bus type
to be obtained from the devicetree.

This fixes the color coding and sync signals were not being read from the devicetree.

Signed-off-by: Mike Looijmans <mike.looijmans@topic.nl>