clk-si5341: Support NVM programming through sysfs
Export an attribute program_nvm_bank that when read reports the current
bank value. To program the chip's current state into NVM, write the
magic value 0xC7 into this attribute.

This allows the clock chip to be programmed "in system" to reduce boot
time by 300ms and allows the clock to be up and running before the
kernel boots (e.g. for bootloader usage).

Signed-off-by: Mike Looijmans <mike.looijmans@topic.nl>