phy-zynqmp: Postpone getting clock rate until it's actually needed
At probe time the driver would display the following error and abort:
  xilinx-psgtr fd400000.phy: Invalid rate 0 for reference clock 0

This issue was that at probe time, the system has not decided yet whether the
GTR is to be used for SATA (150MHz) or PCIe (100 MHz). The driver doesn't need
to know the clock frequency at that point yet, so wait until the lane is
actually being initialized before requesting the clock rate setting.

Signed-off-by: Mike Looijmans <mike.looijmans@topic.nl>